SUREFIRE-762 Add option to include test runtime only dependencies
The 'additionalProfiles'(List<String>) configuration option is used
to recalculate the TestClasspath to include the dependencies from
the given profile ids.

This can be used to include additional dependencies during test
runtime only vs compile/runtime as the normal test scope does.

The implementation rely on resolving Profiles instead of direct
Dependencies to allow IDEs to activate the same Profile ids and
resolve the same Classpath with out any direct support for this
feature in Surefire.

Usage example:

<profile>
	<id>runtime</id>
	<dependencies>
		<dependency>
			...
		</dependency>
	</dependencies>
</profile>

<plugin>
	<groupId>org.apache.maven.plugins</groupId>
	<artifactId>maven-surefire-plugin</artifactId>
	<configuration>
		<additionalProfiles>
			<dependency>runtime</dependency>
		</additionalProfiles>
	</configuration>
</plugin>